+++ changelog.md
@@ -2,6 +2,7 @@ TGUI 0.9.4  (TBD)
 -----------------
 
 - Default theme no longer affects loading widgets from file
+- Fixed issue with renderer property within custom renderer outside tgui namespace
 
 
 TGUI 0.9.3  (15 January 2022)